[CHORE/#30] 액션리스트 뷰 폴더링 수정
