+++ app/src/main/java/com/growthook/aos/presentation/actionlist/proceeding/ProceedingActionlistFragment.kt
@@ -1,18 +1,18 @@
-package com.growthook.aos.presentation.actionlist
+package com.growthook.aos.presentation.actionlist.proceeding
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.growthook.aos.databinding.FragmentCompletedActionlistBinding
+import com.growthook.aos.databinding.FragmentProceedingActionlistBinding
 import com.growthook.aos.util.base.BaseFragment
 
-class CompletedActionlistFragment : BaseFragment<FragmentCompletedActionlistBinding>() {
+class ProceedingActionlistFragment : BaseFragment<FragmentProceedingActionlistBinding>() {
     override fun getFragmentBinding(
         inflater: LayoutInflater,
         container: ViewGroup?,
-    ): FragmentCompletedActionlistBinding =
-        FragmentCompletedActionlistBinding.inflate(inflater, container, false)
+    ): FragmentProceedingActionlistBinding =
+        FragmentProceedingActionlistBinding.inflate(inflater, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
