+++ app/src/main/res/layout/fragment_completed_actionlist.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.actionlist.CompletedActionlistFragment">
+    tools:context=".presentation.actionlist.completed.CompletedActionlistFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
