+++ app/src/main/res/layout/fragment_proceeding_actionlist.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.actionlist.ProceedingActionlistFragment">
+    tools:context=".presentation.actionlist.proceeding.ProceedingActionlistFragment">
 
     <ImageView
         android:id="@+id/iv_proceeding_actionlist_scrap"
